refactor(supernova): use specific bound hooks for fetching and holding filtered
